Reorganised a bit the files
